Add asset to treeview when creating
